Fixing compiler warnings
git-svn-id: https://svn.code.sf.net/p/dansguardian/code/trunk@381 c6e6bffc-5bc6-4360-8735-ace3d95093af

+++ dansguardian/src/authplugins/ip.cpp
@@ -60,7 +60,7 @@ public:
 		addr = a;
 		group = g;
 	};
-	int addr;
+	unsigned long int addr;
 	int group;
 	int operator < (const ip &a) const {
 		return addr < a.addr;
